Extract `readHandicapFromFile()` function into separate utility module
